backend/requirements.txt
Flask>=2.3.0,<4.0.0
Flask-SocketIO>=5.0.0,<6.0.0
Flask-CORS>=4.0.0,<5.0.0
pymongo>=4.0.0,<5.0.0
celery>=5.0.0,<6.0.0
redis>=4.0.0,<6.0.0
requests>=2.25.0,<3.0.0
httpx>=0.20.0,<1.0.0
google-generativeai>=0.2.0,<1.0.0
google-cloud-texttospeech>=2.10.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0
pydantic>=1.8.0,<3.0.0
python-dateutil>=2.8.0,<3.0.0
